+++ kernels/common/scene_geometry_instance.h
@@ -34,7 +34,7 @@ namespace embree
     virtual void enabling ();
     virtual void disabling();
     virtual void setMask (unsigned mask);
-    virtual void setTransform(const AffineSpace3fa& local2world, size_t timeStep);
+    virtual void setTransform(const AffineSpace3fa& local2world, unsigned int timeStep);
     
     void count(const Ref<Geometry>& geom, ssize_t f);
   public:
